Integrate run-server.sh into the Dockerfile
