+++ pyproject.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "flask (>=3.1.2,<4.0.0)",
-    "tna-frontend-jinja (>=0.24.0,<0.25.0)",
+    "tna-frontend-jinja (==0.25.0)",
     "jinja2 (>=3.1.6,<4.0.0)",
     "boto3 (>=1.40.30,<2.0.0)",
     "ghapi (>=1.0.6,<2.0.0)",
